% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{binary_phylogeny}
\alias{binary_phylogeny}
\title{Propose a "phylogenetically"-linked dataset clustering observations based on similarity}
\usage{
binary_phylogeny(data, method = "clustering")
}
\arguments{
\item{method}{Character (default "clustering"): method to cluster observations. Either "clustering" (hierarchical clustering with Manhattan distance) or "nj" (neihgbour joining)}

\item{mat}{A required matrix or data.frame containing binary observations}
}
\value{
A named list with a (possibly reduced) dataset and a putative tree
}
\description{
Propose a "phylogenetically"-linked dataset clustering observations based on similarity
}
\examples{
data = matrix(c(0,0,1, 1,0,0, 1,0,0, 1,0,0), ncol=3, nrow=4)
binary_phylogeny(data)
}
