% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bootstrap_mk_fit}
\alias{bootstrap_mk_fit}
\title{Parameteric bootstrap for a HyperMk model fit}
\usage{
bootstrap_mk_fit(fit, n_boot = 10, reversible = TRUE, flux.samples = 10000)
}
\arguments{
\item{fit}{A fitted HyperMk model (output from hyperinf)}

\item{n_boot}{Numeric (default 10) number of bootstrap resamples to run}

\item{reversible}{Boolean (default TRUE) whether we're running a reversible HyperMk model}

\item{flux.samples}{Numeric (default 10000) number of samplers to run to characterise dynamics}
}
\value{
The original fitted model with a new element boots, containing transition matrices and dynamic summaries across bootstraps
}
\description{
Parameteric bootstrap for a HyperMk model fit
}
