% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperinf.R
\name{clean_data}
\alias{clean_data}
\title{Curate data into a systematic format for inference}
\usage{
clean_data(data)
}
\arguments{
\item{data}{A required matrix or data.frame containing binary observations}
}
\value{
A matrix of observations in a standardised format
}
\description{
Data can be flexibly supplied as a matrix or data frame, with or without a first labelling column, and with binary observations 0, 1 (with ? or 2 for uncertainty) as individual columns or concatenated strings
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
mat = clean_data(data)
}
