% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-model.R
\name{full_to_squared}
\alias{full_to_squared}
\title{Converts a fully parameterised model to a set of parameters for L^2 HyperTraPS}
\usage{
full_to_squared(fit)
}
\arguments{
\item{fit}{A fitted model with a fully parameterised hypercube (may include bootstrap resamples)}
}
\value{
A vector of L^2 values corresponding to the parameterisation of a HyperTraPS model
}
\description{
Converts a fully parameterised model to a set of parameters for L^2 HyperTraPS
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data)
l2params = full_to_squared(fit)
}
