% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-model.R
\name{full_to_squared_fit}
\alias{full_to_squared_fit}
\title{Estimates an L^2 HyperTraPS model from fully parameterised HyperHMM/LAU/Mk model}
\usage{
full_to_squared_fit(fit, reps = 100)
}
\arguments{
\item{fit}{A fitted model from HyperHMM, HyperLAU, or (irreversible) HyperMk}

\item{reps}{An integer (default 100) of times to repeat this parameterisation to build up a model structure}
}
\value{
A HyperTraPS model structure
}
\description{
Combines full_to_squared and hypertraps_from_params
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data)
est.ht = full_to_squared_fit(fit)
}
