% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_plot_graph}
\alias{get_plot_graph}
\title{Get a graph object reflecting a (plottable) transition network from a fitted model}
\usage{
get_plot_graph(
  fit,
  fit.type = NULL,
  uncertainty = FALSE,
  reversible = FALSE,
  threshold = 0.05,
  feature.names = NULL,
  max.samps = 1000
)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{fit.type}{(default NULL) character vector giving type of fit (overridden by inferred type)}

\item{uncertainty}{Boolean, are we bootstrapping}

\item{reversible}{Boolean, are we consider reversible steps}

\item{threshold}{Numeric (default 0.05), threshold to place on fluxes}

\item{feature.names}{NULL or character vector (default NULL). If vector of length L, use those labels.}

\item{max.samps}{Numeric (default 1000) samples for HyperTraPS sampling}
}
\value{
A graph object
}
\description{
Get a graph object reflecting a (plottable) transition network from a fitted model
}
