% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{hyperinf_AIC}
\alias{hyperinf_AIC}
\title{Get the AIC for a fitted hypercubic inference model}
\usage{
hyperinf_AIC(fit, ...)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{...}{other parameters to pass to hyperinf_loglikelihood}
}
\value{
A numerical value
}
\description{
Get the AIC for a fitted hypercubic inference model
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data)
hyperinf_AIC(fit)
}
