% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative.R
\name{hyperinf_bubbles}
\alias{hyperinf_bubbles}
\title{Produce dataframe of ordering probabilities}
\usage{
hyperinf_bubbles(fit)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}
}
\value{
A dataframe with columns feature/order/prob
}
\description{
Produce dataframe of ordering probabilities
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit.1 = hyperinf(data)
hyperinf_bubbles(fit.1)
}
