% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hyperinf_diagnostics}
\alias{hyperinf_diagnostics}
\title{Run diagnostics on a model fit}
\usage{
hyperinf_diagnostics(fit, show.plot = TRUE)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{show.plot}{Boolean (default TRUE) whether to plot a diagnostic plot if appropriate}
}
\value{
A dataframe with descriptive statistics depending on fit type. Messages are printed to console if warning signs exist for model convergence.
}
\description{
Run diagnostics on a model fit
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit.1 = hyperinf(data, method="hypertraps")
hyperinf_diagnostics(fit.1)
}
