% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{hyperinf_estimate_regularised}
\alias{hyperinf_estimate_regularised}
\title{Crude regularisation estimate of a hypercubic inference model}
\usage{
hyperinf_estimate_regularised(fit, threshold = 0.001)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{threshold}{Numeric (default 1e-3), threshold of probability flux below which to remove a transition}
}
\value{
A list of regularised model and AIC statistics before and after "regularisation"
}
\description{
Crude regularisation estimate of a hypercubic inference model
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data, method="hyperhmm")
hyperinf_estimate_regularised(fit)
}
