% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hyperinf_gettype}
\alias{hyperinf_gettype}
\title{Get the fitting method used to produce a hyperinf model fit}
\usage{
hyperinf_gettype(fit)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}
}
\value{
A character string describing the fit type
}
\description{
Get the fitting method used to produce a hyperinf model fit
}
