% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{hyperinf_loglikelihood}
\alias{hyperinf_loglikelihood}
\title{Get the likelihood for a fitted hypercubic inference model}
\usage{
hyperinf_loglikelihood(fit, emission = FALSE, expanded = FALSE)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{emission}{Boolean (default FALSE), whether to consider emission probabilities P(s->t)Pemit(s, t) or just P(s->t), the probability that a random walker passes through s and t}

\item{expanded}{Boolean (default FALSE) for HyperTraPS, whether to report the range of sampled likelihoods or just the max of this sample}
}
\value{
A numerical value or (for expanded HyperTraPS) a vector of values
}
\description{
Get the likelihood for a fitted hypercubic inference model
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data)
hyperinf_loglikelihood(fit)
}
