% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{hyperinf_regularise}
\alias{hyperinf_regularise}
\title{Regularisation of a hypercubic inference model}
\usage{
hyperinf_regularise(fit, threshold = 0)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{threshold}{Numeric (default 0), threshold of probability flux below which to remove a transition}
}
\value{
A list of regularised model and AIC statistics before and after "regularisation"
}
\description{
Regularisation of a hypercubic inference model
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data, method="hypermk")
hyperinf_regularise(fit)
}
