% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{hyperinf_span_relatedness}
\alias{hyperinf_span_relatedness}
\title{Compare inference outputs assuming cross-sectional or minimally convergent dynamics}
\usage{
hyperinf_span_relatedness(mat, method = "clustering", ...)
}
\arguments{
\item{mat}{A required matrix or data.frame containing binary observations}

\item{method}{Character (default "clustering"): method to cluster observations. Either "clustering" (hierarchical clustering with Manhattan distance) or "nj" (neihgbour joining)}

\item{...}{Other arguments to pass to hyperinf}
}
\value{
A ggplot object
}
\description{
Compare inference outputs assuming cross-sectional or minimally convergent dynamics
}
\examples{
data = matrix(c(0,0,1, 1,0,0, 1,0,0, 1,0,0), ncol=3, nrow=4)
hyperinf_span_relatedness(data)
}
