% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_matrix.R
\name{multiple_fits_to_booted_fit}
\alias{multiple_fits_to_booted_fit}
\title{Case multiple model fits into a form labelled as bootstrap samples (for compatibility)}
\usage{
multiple_fits_to_booted_fit(fits)
}
\arguments{
\item{fits}{A list of fitted hypercubic inference models (output from hyperinf)}
}
\value{
A hypercubic inference model structure with the model fits stored as "bootstrap resamples"
}
\description{
Case multiple model fits into a form labelled as bootstrap samples (for compatibility)
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit.1 = hyperinf(data, method="hypertraps", seed=1)
fit.2 = hyperinf(data, method="hypertraps", seed=2)
fit = multiple_fits_to_booted_fit(list(fit.1, fit.2))
}
