% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_matrix.R
\name{ordering_matrix}
\alias{ordering_matrix}
\title{Compute a feature ordering matrix for a hypercubic inference model fit}
\usage{
ordering_matrix(fit, n.samples = 10000, type = "relative")
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{n.samples}{Integer (default 10k) number of samples to characterise dynamics in the reversible (HyperMk) case}

\item{type}{Character string (default "relative"). Either "relative", in which case P_ij gives the probability that feature i is acquired before feature j. Or "absolute", in which case P_ij is the probability that feature i is acquired after step j.}
}
\value{
A matrix giving P_ij according to the type of comparison (see above)
}
\description{
Compute a feature ordering matrix for a hypercubic inference model fit
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit.1 = hyperinf(data)
ordering_matrix(fit.1)
ordering_matrix(fit.1, type="absolute")
}
