% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperinf.R
\name{plot_hyperinf}
\alias{plot_hyperinf}
\title{Plot a fitted hypercubic inference model}
\usage{
plot_hyperinf(fit, plot.type = "", threshold = 0.05, uncertainty = "")
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{plot.type}{A character string, either empty (default) to allow standardised plot, or "native" to produce plots from the source algorithm}

\item{threshold}{Double (default 0.05), probability flux below which edges will not be plotted}

\item{uncertainty}{Boolean, whether to visualise uncertainty over bootstraps (only for HyperLAU and HyperHMM)}
}
\value{
A ggplot object
}
\description{
Plot a fitted hypercubic inference model
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data)
plot_hyperinf(fit)
}
