% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative.R
\name{plot_hyperinf_bootstrap}
\alias{plot_hyperinf_bootstrap}
\title{Comparative bubble plot for bootstrapped fitted models}
\usage{
plot_hyperinf_bootstrap(fit.1, fit.2)
}
\arguments{
\item{fit.1}{A fitted hypercubic inference models with bootstrap info (output from hyperinf)}

\item{fit.2}{A fitted hypercubic inference models with bootstrap info (output from hyperinf)}
}
\value{
A ggplot object
}
\description{
Comparative bubble plot for bootstrapped fitted models
}
