% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative.R
\name{plot_hyperinf_bubbles}
\alias{plot_hyperinf_bubbles}
\title{Compare bubble plots for a set of fitted hypercubic inference models}
\usage{
plot_hyperinf_bubbles(
  fits,
  reorder = FALSE,
  transpose = FALSE,
  thetastep = 10,
  p.scale = 1,
  sqrt.trans = FALSE,
  bins = 0,
  expt.names = NULL,
  fill.name = "Experiment",
  feature.names = TRUE
)
}
\arguments{
\item{fits}{A list of fitted hypercubic inference models (output from hyperinf)}

\item{reorder}{Boolean (default FALSE), whether to reorder features by mean acquisition order}

\item{transpose}{Boolean (default FALSE), whether to transpose axes for the bubble plot}

\item{thetastep}{Integer (default 10), the number of angular steps to take when drawing polygons for each bubble segment. Lower this for many comparisons.}

\item{p.scale}{Numeric (default 1), the scaling factor to apply to the radius (probability) of bubbles}

\item{sqrt.trans}{Boolean (default FALSE), whether to sqrt-transform the probability to get the bubble radius}

\item{bins}{Integer (default 0, off), how many bins to divide the ordering axis into}

\item{expt.names}{Optional vector of labels for each element of the fit list}

\item{fill.name}{Character (default "Experiment"), the label to use for the group type}

\item{feature.names}{Boolean or character vector (default TRUE). If TRUE, use feature names from fit. If FALSE, use numerical labels. If vector of length L, use those labels.}
}
\value{
A ggplot object
}
\description{
Compare bubble plots for a set of fitted hypercubic inference models
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit.1 = hyperinf(data)
fit.2 = hyperinf(1-data)
plot_hyperinf_bubbles(list(fit.1, fit.2))
}
