% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative.R
\name{plot_hyperinf_comparative}
\alias{plot_hyperinf_comparative}
\title{Compare transition graphs for a set of fitted hypercubic inference models}
\usage{
plot_hyperinf_comparative(
  fits,
  threshold = 0.05,
  expt.names = NULL,
  feature.names = TRUE,
  style = "limited",
  bend = 0.5,
  label_size = 2
)
}
\arguments{
\item{fits}{A list of fitted hypercubic inference models (output from hyperinf)}

\item{threshold}{Double (default 0.05), probability flux below which edges will not be plotted}

\item{expt.names}{Optional vector of labels for each element of the fit list}

\item{feature.names}{Boolean or character vector (default TRUE). If TRUE, use feature names from fit. If FALSE, use numerical labels. If vector of length L, use those labels.}

\item{style}{Character (default "limited") giving plot style. "limited" condenses bootstrap resamples into single edges; others plot each as a different arc.}

\item{bend}{Numeric (default 0.5) the strength of the bend separating edges between the same pair of nodes}

\item{label_size}{Numeric (default 2) the size of edge labels}
}
\value{
A ggplot object
}
\description{
Compare transition graphs for a set of fitted hypercubic inference models
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit.1 = hyperinf(data)
fit.2 = hyperinf(1-data)
plot_hyperinf_comparative(list(fit.1, fit.2))
}
