% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_matrix.R
\name{plot_hyperinf_compare_orderings}
\alias{plot_hyperinf_compare_orderings}
\title{Plot comparative ordering statistics from two fitted models}
\usage{
plot_hyperinf_compare_orderings(fit.1, fit.2, threshold = 0.2, ...)
}
\arguments{
\item{fit.1}{A fitted hypercubic inference model including bootstrap resamples (output from hyperinf)}

\item{fit.2}{Another fitted hypercubic inference model including bootstrap resamples  (output from hyperinf)}

\item{threshold}{Numeric (default 0). A threshold for differences. If zero, any separation of statistic (e.g. non-overlapping bootstrap distributions) is reported. If nonzero, we require one fit's statistic > 1-threshold and the other fit's statistic < threshold.}

\item{...}{additional parameters to pass to plot_hyperinf_bubbles}
}
\value{
A ggplot object displaying pairs of features and timings for which separation is detected
}
\description{
Plot comparative ordering statistics from two fitted models
}
