% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperinf.R
\name{plot_hyperinf_interactions}
\alias{plot_hyperinf_interactions}
\title{Plot interactions from a fitted hypercubic inference model}
\usage{
plot_hyperinf_interactions(fit, threshold = 0.1, cv.threshold = 0.5)
}
\arguments{
\item{fit}{A fitted hypercubic inference model (output from hyperinf)}

\item{threshold}{Double (default 0.05), interaction strength below which edges will not be plotted}

\item{cv.threshold}{Double (default 1), coefficient of variation for interaction strength above which edges will not be plotted}
}
\value{
A ggplot object
}
\description{
Plot interactions from a fitted hypercubic inference model
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data)
plot_hyperinf_interactions(fit)
}
