% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_matrix.R
\name{plot_hyperinf_ordering_matrices}
\alias{plot_hyperinf_ordering_matrices}
\title{Plot ordering matrices from fitted models}
\usage{
plot_hyperinf_ordering_matrices(
  fits,
  expt.names = NULL,
  feature.names = TRUE,
  thetastep = 5,
  p.scale = 0.5,
  sqrt.trans = FALSE
)
}
\arguments{
\item{fits}{A list of fitted hypercubic inference models (or single model)}

\item{expt.names}{Optional vector of labels for each element of the fit list}

\item{feature.names}{Boolean or character vector (default TRUE). If TRUE, use feature names from fit. If FALSE, use numerical labels. If vector of length L, use those labels.}

\item{thetastep}{Integer (default 5), the number of angular steps to take when drawing polygons for each bubble segment. Lower this for many comparisons.}

\item{p.scale}{Numeric (default 1), the scaling factor to apply to the radius (probability) of bubbles}

\item{sqrt.trans}{Boolean (default FALSE), whether to sqrt-transform the probability to get the bubble radius}
}
\value{
A ggplot object comparing ordering matrices
}
\description{
Plot ordering matrices from fitted models
}
