% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_hyperinf_pca}
\alias{plot_hyperinf_pca}
\title{Plot PCA across fitted models}
\usage{
plot_hyperinf_pca(fits, expt.names = NULL)
}
\arguments{
\item{fits}{A list of fitted hypercubic inference models (output from hyperinf)}

\item{expt.names}{Optional vector of labels for each element of the fit list}
}
\value{
A ggplot object
}
\description{
Plot PCA across fitted models
}
