% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_matrix.R
\name{compare_orderings}
\alias{compare_orderings}
\title{Compare ordering statistics from two fitted models}
\usage{
compare_orderings(
  fit.1,
  fit.2,
  n.samples = 10000,
  threshold = 0,
  type = "relative"
)
}
\arguments{
\item{fit.1}{A fitted hypercubic inference model including bootstrap resamples (output from hyperinf)}

\item{fit.2}{Another fitted hypercubic inference model including bootstrap resamples  (output from hyperinf)}

\item{n.samples}{Integer (default 10k) number of samples to characterise dynamics in the reversible (HyperMk) case}

\item{threshold}{Numeric (default 0). A threshold for differences. If zero, any separation of statistic (e.g. non-overlapping bootstrap distributions) is reported. If nonzero, we require one fit's statistic > 1-threshold and the other fit's statistic < threshold.}

\item{type}{Character string (default "relative"). Either "relative", in which case P_ij gives the probability that feature i is acquired before feature j. Or "absolute", in which case P_ij is the probability that feature i is acquired after step j.}
}
\value{
A hypercubic inference model structure with the model fits stored as "bootstrap resamples"
}
\description{
Compare ordering statistics from two fitted models
}
