% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperinf.R
\name{hyperinf}
\alias{hyperinf}
\title{Do hypercubic inference on some data}
\usage{
hyperinf(
  data,
  tree = NULL,
  losses = FALSE,
  method = "",
  reversible = FALSE,
  boot.parallel = 0,
  ...
)
}
\arguments{
\item{data}{A required matrix or data.frame containing binary observations}

\item{tree}{Optional tree object}

\item{losses}{Boolean (default FALSE) whether to consider losses rather than gains of features}

\item{method}{A character string, either empty (default) to allow automatic choice of algorithm, or one of the options in the text above}

\item{reversible}{Boolean (default FALSE) whether to allow reversible transitions}

\item{boot.parallel}{Integer (default 0) number of bootstrap resamples to run in parallel (only meaningful for HyperHMM and HyperLAU)}

\item{...}{other options to pass to the inference method used. For example, nboot for HyperHMM, length/kernel/walkers for HyperTraPS.}
}
\value{
A fitted hypercubic inference object
}
\description{
Data can be flexibly supplied as a matrix or data frame, with or without an accompanying phylogeny
describing how observations are related
}
\details{
Method can be specified: "hypermk", "hyperhmm", "hypertraps", "pli", or "hyperdags". If unspecified, the
most detailed approach compatible with the data will be chosen
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = hyperinf(data)
}
