% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative.R
\name{plot_hyperinf_comparative}
\alias{plot_hyperinf_comparative}
\title{Compare transition graphs for a set of fitted hypercubic inference models}
\usage{
plot_hyperinf_comparative(
  fits,
  threshold = 0.05,
  expt.names = NULL,
  feature.names = NULL
)
}
\arguments{
\item{fits}{A list of fitted hypercubic inference models (output from hyperinf)}

\item{threshold}{Double (default 0.05), probability flux below which edges will not be plotted}

\item{expt.names}{Optional vector of labels for each element of the fit list}

\item{feature.names}{Optional vector of labels for the L features in each fit}
}
\value{
A ggplot object
}
\description{
Compare transition graphs for a set of fitted hypercubic inference models
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit.1 = hyperinf(data)
fit.2 = hyperinf(1-data)
plot_hyperinf_comparative(list(fit.1, fit.2))
}
