% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperinf.R
\name{plot_hyperinf_data}
\alias{plot_hyperinf_data}
\title{Plot some data for accumulation modelling}
\usage{
plot_hyperinf_data(
  data,
  tree = NULL,
  feature.names = TRUE,
  hjust = 1,
  bmargin = 40,
  ...
)
}
\arguments{
\item{data}{A required matrix or data.frame containing binary observations}

\item{tree}{Optional tree object}

\item{hjust}{Numeric (default 1), horizontal justification for rotated feature labels}

\item{bmargin}{Numeric (default 40), bottom margin size to prevent labels being truncated}

\item{...}{other options to pass to plotHypercube.curated.tree (if tree is provided)}
}
\value{
A ggplot object
}
\description{
Plot some data for accumulation modelling
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
plot_hyperinf_data(data)
}
